refactoring, cleaning up exposed SDK
