+++ cmd/maestro/run/run.go
@@ -9,9 +9,9 @@ import (
 	"github.com/jexia/maestro/cmd/maestro/config"
 	"github.com/jexia/maestro/codec/json"
 	"github.com/jexia/maestro/codec/proto"
-	"github.com/jexia/maestro/constructor"
 	"github.com/jexia/maestro/definitions/hcl"
-	"github.com/jexia/maestro/logger"
+	"github.com/jexia/maestro/internal/constructor"
+	"github.com/jexia/maestro/internal/logger"
 	"github.com/jexia/maestro/schema/protoc"
 	"github.com/jexia/maestro/specs"
 	"github.com/jexia/maestro/transport/graphql"
