+++ cmd/maestro/validate/validate.go
@@ -5,10 +5,10 @@ import (
 	"github.com/jexia/maestro/cmd/maestro/config"
 	"github.com/jexia/maestro/codec/json"
 	"github.com/jexia/maestro/codec/proto"
-	"github.com/jexia/maestro/constructor"
 	"github.com/jexia/maestro/definitions/hcl"
-	"github.com/jexia/maestro/instance"
-	"github.com/jexia/maestro/logger"
+	"github.com/jexia/maestro/internal/constructor"
+	"github.com/jexia/maestro/internal/instance"
+	"github.com/jexia/maestro/internal/logger"
 	"github.com/jexia/maestro/schema/protoc"
 	"github.com/jexia/maestro/transport/grpc"
 	"github.com/jexia/maestro/transport/http"
