+++ definitions/hcl/specs_test.go
@@ -6,8 +6,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/jexia/maestro/instance"
-	"github.com/jexia/maestro/utils"
+	"github.com/jexia/maestro/internal/instance"
+	"github.com/jexia/maestro/internal/utils"
 )
 
 const (
