+++ examples/http/main.go
@@ -5,7 +5,7 @@ import (
 	"github.com/jexia/maestro/codec/json"
 	"github.com/jexia/maestro/codec/proto"
 	"github.com/jexia/maestro/definitions/hcl"
-	"github.com/jexia/maestro/logger"
+	"github.com/jexia/maestro/internal/logger"
 	"github.com/jexia/maestro/schema/protoc"
 	"github.com/jexia/maestro/specs"
 	"github.com/jexia/maestro/transport/http"
