+++ internal/constructor/constructor.go
@@ -2,8 +2,8 @@ package constructor
 
 import (
 	"github.com/jexia/maestro/codec"
-	"github.com/jexia/maestro/flow"
-	"github.com/jexia/maestro/instance"
+	"github.com/jexia/maestro/internal/flow"
+	"github.com/jexia/maestro/internal/instance"
 	"github.com/jexia/maestro/metadata"
 	"github.com/jexia/maestro/schema"
 	"github.com/jexia/maestro/specs"
