+++ internal/constructor/options.go
@@ -2,7 +2,7 @@ package constructor
 
 import (
 	"github.com/jexia/maestro/codec"
-	"github.com/jexia/maestro/instance"
+	"github.com/jexia/maestro/internal/instance"
 	"github.com/jexia/maestro/schema"
 	"github.com/jexia/maestro/specs"
 	"github.com/jexia/maestro/transport"
