+++ internal/flow/flow.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"sync"
 
-	"github.com/jexia/maestro/instance"
-	"github.com/jexia/maestro/logger"
+	"github.com/jexia/maestro/internal/instance"
+	"github.com/jexia/maestro/internal/logger"
 	"github.com/jexia/maestro/specs"
 	"github.com/sirupsen/logrus"
 )
