+++ internal/flow/node.go
@@ -3,8 +3,8 @@ package flow
 import (
 	"context"
 
-	"github.com/jexia/maestro/instance"
-	"github.com/jexia/maestro/logger"
+	"github.com/jexia/maestro/internal/instance"
+	"github.com/jexia/maestro/internal/logger"
 	"github.com/jexia/maestro/specs"
 	"github.com/sirupsen/logrus"
 )
