+++ main_test.go
@@ -6,14 +6,14 @@ import (
 	"time"
 
 	"github.com/jexia/maestro/codec/json"
-	"github.com/jexia/maestro/constructor"
 	"github.com/jexia/maestro/definitions/hcl"
-	"github.com/jexia/maestro/instance"
-	"github.com/jexia/maestro/logger"
+	"github.com/jexia/maestro/internal/constructor"
+	"github.com/jexia/maestro/internal/instance"
+	"github.com/jexia/maestro/internal/logger"
+	"github.com/jexia/maestro/internal/utils"
 	"github.com/jexia/maestro/schema/mock"
 	"github.com/jexia/maestro/specs"
 	"github.com/jexia/maestro/transport/http"
-	"github.com/jexia/maestro/utils"
 )
 
 func TestOptions(t *testing.T) {
