+++ options.go
@@ -2,9 +2,9 @@ package maestro
 
 import (
 	"github.com/jexia/maestro/codec"
-	"github.com/jexia/maestro/constructor"
-	"github.com/jexia/maestro/instance"
-	"github.com/jexia/maestro/logger"
+	"github.com/jexia/maestro/internal/constructor"
+	"github.com/jexia/maestro/internal/instance"
+	"github.com/jexia/maestro/internal/logger"
 	"github.com/jexia/maestro/schema"
 	"github.com/jexia/maestro/specs"
 	"github.com/jexia/maestro/transport"
