+++ schema/schema.go
@@ -1,7 +1,7 @@
 package schema
 
 import (
-	"github.com/jexia/maestro/instance"
+	"github.com/jexia/maestro/internal/instance"
 	"github.com/jexia/maestro/specs/labels"
 	"github.com/jexia/maestro/specs/types"
 )
