+++ schema/store.go
@@ -1,8 +1,8 @@
 package schema
 
 import (
-	"github.com/jexia/maestro/instance"
-	"github.com/jexia/maestro/logger"
+	"github.com/jexia/maestro/internal/instance"
+	"github.com/jexia/maestro/internal/logger"
 )
 
 // NewStore constructs a new schema store
