+++ specs/dependencies/references_test.go
@@ -3,7 +3,7 @@ package dependencies
 import (
 	"testing"
 
-	"github.com/jexia/maestro/instance"
+	"github.com/jexia/maestro/internal/instance"
 	"github.com/jexia/maestro/specs"
 )
