+++ specs/strict/checks.go
@@ -3,8 +3,8 @@ package strict
 import (
 	"sync"
 
-	"github.com/jexia/maestro/instance"
-	"github.com/jexia/maestro/logger"
+	"github.com/jexia/maestro/internal/instance"
+	"github.com/jexia/maestro/internal/logger"
 	"github.com/jexia/maestro/specs"
 	"github.com/jexia/maestro/specs/trace"
 )
