+++ transport/grpc/caller.go
@@ -8,8 +8,8 @@ import (
 
 	"google.golang.org/grpc/metadata"
 
-	"github.com/jexia/maestro/instance"
-	"github.com/jexia/maestro/logger"
+	"github.com/jexia/maestro/internal/instance"
+	"github.com/jexia/maestro/internal/logger"
 	"github.com/jexia/maestro/schema"
 	"github.com/jexia/maestro/specs"
 	"github.com/jexia/maestro/specs/trace"
