+++ transport/grpc/caller_test.go
@@ -7,8 +7,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/jexia/maestro/flow"
-	"github.com/jexia/maestro/instance"
+	"github.com/jexia/maestro/internal/flow"
+	"github.com/jexia/maestro/internal/instance"
 	"github.com/jexia/maestro/metadata"
 	"github.com/jexia/maestro/schema"
 	"github.com/jexia/maestro/schema/mock"
