+++ transport/grpc/grpc_test.go
@@ -7,8 +7,8 @@ import (
 	"testing"
 
 	"github.com/jexia/maestro/codec"
-	"github.com/jexia/maestro/flow"
-	"github.com/jexia/maestro/instance"
+	"github.com/jexia/maestro/internal/flow"
+	"github.com/jexia/maestro/internal/instance"
 	"github.com/jexia/maestro/schema/mock"
 	"github.com/jexia/maestro/specs"
 	"github.com/jexia/maestro/specs/labels"
