+++ transport/grpc/listener.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/jexia/maestro/codec"
 	"github.com/jexia/maestro/codec/proto"
-	"github.com/jexia/maestro/instance"
-	"github.com/jexia/maestro/logger"
+	"github.com/jexia/maestro/internal/instance"
+	"github.com/jexia/maestro/internal/logger"
 	"github.com/jexia/maestro/metadata"
 	"github.com/jexia/maestro/specs"
 	"github.com/jexia/maestro/transport"
