+++ transport/http/caller_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 
 	"github.com/jexia/maestro/codec/json"
-	"github.com/jexia/maestro/instance"
+	"github.com/jexia/maestro/internal/instance"
 	"github.com/jexia/maestro/metadata"
 	"github.com/jexia/maestro/specs"
 	"github.com/jexia/maestro/specs/labels"
