+++ transport/http/listener_test.go
@@ -10,8 +10,8 @@ import (
 
 	"github.com/jexia/maestro/codec"
 	"github.com/jexia/maestro/codec/json"
-	"github.com/jexia/maestro/flow"
-	"github.com/jexia/maestro/instance"
+	"github.com/jexia/maestro/internal/flow"
+	"github.com/jexia/maestro/internal/instance"
 	"github.com/jexia/maestro/specs"
 	"github.com/jexia/maestro/transport"
 )
