+++ transport/protocol.go
@@ -5,7 +5,7 @@ import (
 	"io"
 
 	"github.com/jexia/maestro/codec"
-	"github.com/jexia/maestro/instance"
+	"github.com/jexia/maestro/internal/instance"
 	"github.com/jexia/maestro/metadata"
 	"github.com/jexia/maestro/schema"
 	"github.com/jexia/maestro/specs"
